pyproject.toml
[tool.poetry]
name = "aws-assistant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.1"
langchain = "^0.1.20"
uvicorn = ">=0.25.0,<0.26.0"
openai = "^1.28.1"
fastapi = ">=0.110.1,<0.111.0"
langchain-openai = "^0.1.6"
chainlit = "^1.1.0"
tiktoken = "^0.7.0"
boto3 = "^1.34.112"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
